Added RubyMine's .idea to .gitignore. Deleted migrations from evaluations. Fixed UserRole#clean_abstract_role.

+++ app/models/concerns/user_role.rb
@@ -11,7 +11,7 @@ module UserRole
         delegate :full_name, :login_full_name, :login, to: :user
 
         def clean_abstract_role
-            self.abstract_role.destroy
+            self.abstract_role.destroy if self && self.abstract_role
         end
     end
 end
